+++ csc/2016/vsv/src/server/Server.kt
@@ -1,6 +1,6 @@
 package server
 
-import task.RequestExecutorService
+import task.handlers.RequestExecutorService
 import java.io.IOException
 import java.net.ServerSocket
 import java.util.concurrent.atomic.AtomicBoolean
