+++ csc/2016/vsv/src/task/handlers/requests/CalculateRequestExecutor.kt
@@ -1,11 +1,10 @@
-package task.requests
+package task.handlers.requests
 
 import communication.CommunicationProtos
 
 
-class SubscribeRequestExecutor() : RequestExecutor {
+class CalculateRequestExecutor() : RequestExecutor {
     override fun execute(request: CommunicationProtos.ServerRequest): CommunicationProtos.ServerResponse {
         throw UnsupportedOperationException()
     }
-
 }
