+++ csc/2016/vsv/src/task/handlers/requests/SubscribeRequestExecutor.kt
@@ -1,10 +1,11 @@
-package task.requests
+package task.handlers.requests
 
 import communication.CommunicationProtos
 
 
-class CalculateRequestExecutor() : RequestExecutor {
+class SubscribeRequestExecutor() : RequestExecutor {
     override fun execute(request: CommunicationProtos.ServerRequest): CommunicationProtos.ServerResponse {
         throw UnsupportedOperationException()
     }
+
 }
